refactor: move `DataPointRepository` into the influx package
