+++ src/main/kotlin/com/github/shirleh/datacollection/BanDataCollector.kt
@@ -1,5 +1,6 @@
 package com.github.shirleh.datacollection
 
+import com.github.shirleh.persistence.influx.DataPointRepository
 import com.influxdb.client.domain.WritePrecision
 import com.influxdb.client.write.Point
 import discord4j.common.util.Snowflake
