+++ src/main/kotlin/com/github/shirleh/datacollection/MemberLeaveDataCollector.kt
@@ -1,6 +1,7 @@
 package com.github.shirleh.datacollection
 
 import com.github.shirleh.extensions.orElseNull
+import com.github.shirleh.persistence.influx.DataPointRepository
 import com.influxdb.client.domain.WritePrecision
 import com.influxdb.client.write.Point
 import discord4j.core.event.domain.guild.MemberLeaveEvent
