--- src/main/kotlin/com/github/shirleh/datacollection/Module.kt
@@ -1,7 +1,0 @@
-package com.github.shirleh.datacollection
-
-import org.koin.dsl.module
-
-val dataCollectionModule = module {
-    single<DataPointRepository> { DataPointRepositoryImpl(get()) }
-}
