+++ src/main/kotlin/com/github/shirleh/persistence/influx/Module.kt
@@ -4,7 +4,9 @@ import com.influxdb.client.InfluxDBClient
 import org.koin.dsl.module
 
 val influxModule = module {
-    single { InfluxConfiguration(get()) }
+    single(createdAtStart = true) { InfluxConfiguration(get()) }
     single { InfluxClientFactory.create(get()) }
     single { get<InfluxDBClient>().writeApi }
+
+    single<DataPointRepository> { DataPointRepositoryImpl(get()) }
 }
